switch from file names with issue numbers to file names with "old"
