+++ web-client/src/presenter/actions/TrialSession/setJudgesCaseNoteOnCaseDetailAction.old.js
@@ -1,12 +1,12 @@
 import { state } from 'cerebral';
 
 /**
- * sets the state.judgesNote
+ * sets the state.caseDetail.judgesNote
  *
  * @param {object} providers the providers object
  * @param {object} providers.props the cerebral props object containing the props.userNote
  * @param {object} providers.store the cerebral store used for setting the state
  */
 export const setJudgesCaseNoteOnCaseDetailAction = ({ props, store }) => {
-  store.set(state.judgesNote, props.userNote);
+  store.set(state.caseDetail.judgesNote, props.userNote);
 };
