+++ web-client/src/presenter/actions/TrialSession/setJudgesCaseNoteOnCaseDetailAction.old.test.js
@@ -1,5 +1,5 @@
 import { runAction } from 'cerebral/test';
-import { setJudgesCaseNoteOnCaseDetailAction } from './setJudgesCaseNoteOnCaseDetailAction.6979';
+import { setJudgesCaseNoteOnCaseDetailAction } from './setJudgesCaseNoteOnCaseDetailAction.old';
 
 describe('setJudgesCaseNoteOnCaseDetailAction', () => {
   it('sets default trial session detail tab', async () => {
@@ -9,6 +9,6 @@ describe('setJudgesCaseNoteOnCaseDetailAction', () => {
       },
     });
 
-    expect(result.state.judgesNote).toEqual('welcome to flavortown');
+    expect(result.state.caseDetail.judgesNote).toEqual('welcome to flavortown');
   });
 });
